refactor: extract map runner trace helpers
